# backend/main.go
package main

import (
	"log"
	"os"

	"github.com/gofiber/fiber/v2"
	"github.com/gofiber/fiber/v2/middleware/cors"
	"github.com/gofiber/fiber/v2/middleware/logger"
	"github.com/gofiber/fiber/v2/middleware/requestid"
	"github.com/gofiber/swagger"
	_ "github.com/repoareta/pedeve-dms-app/backend/docs" // docs is generated by Swag CLI
	"github.com/repoareta/pedeve-dms-app/backend/internal/delivery/http"
	"github.com/repoareta/pedeve-dms-app/backend/internal/usecase"
)

// @title           Pedeve App Backend API
// @version         1.0
// @description     Pedeve App Backend API
// @termsOfService  http://swagger.io/terms/

// @contact.name   API Support
// @contact.url    http://www.swagger.io/support
// @contact.email  support@swagger.io

// @license.name  Apache 2.0
// @license.url   http://www.apache.org/licenses/LICENSE-2.0.html

// @host      localhost:8080
// @BasePath  /api/v1

// @securityDefinitions.apikey  BearerAuth
// @in                          header
// @name                        Authorization
// @description                 Type "Bearer" followed by a space and JWT token. Atau menggunakan httpOnly cookie (auth_token) untuk keamanan yang lebih baik.

// @note                        Catatan Teknis API:
// @note                        1. Framework: Menggunakan Fiber v2 dengan performa tinggi (fasthttp)
// @note                        2. Authentication: JWT token dengan httpOnly cookie untuk mencegah XSS attacks
// @note                        3. Rate Limiting:
// @note                           - General API: 100 requests/second, burst: 50
// @note                           - Auth endpoints: 5 requests/minute, burst: 5 (untuk prevent brute force)
// @note                        4. Security Headers: X-Content-Type-Options, X-XSS-Protection, CSP, HSTS
// @note                        5. CSRF Protection: Double-submit cookie pattern untuk state-changing requests (POST, PUT, DELETE, PATCH)
// @note                        6. Audit Logging: Semua aksi user dan error teknis dicatat dalam audit log
// @note                        7. 2FA Support: TOTP-based 2FA dengan backup codes untuk recovery
// @note                        8. Error Handling: Centralized error logging dengan stack trace untuk debugging
// @note                        9. Database: PostgreSQL/SQLite dengan GORM ORM, auto-migration enabled
// @note                        10. CORS: Configured untuk localhost:5173 (Vite dev) dan localhost:3000
func main() {
	// Inisialisasi database
	InitDB()

	// Inisialisasi audit logger
	InitAuditLogger()

	// Mulai cleanup token CSRF
	StartCSRFTokenCleanup()

	// Mulai cleanup audit logs (retention policy)
	StartAuditLogCleanup()

	// Seed user superadmin
	SeedSuperAdmin()

	// Setup Fiber app
	app := fiber.New(fiber.Config{
		ErrorHandler: func(c *fiber.Ctx, err error) error {
			// Custom error handler untuk logging error ke audit log
			code := fiber.StatusInternalServerError
			if e, ok := err.(*fiber.Error); ok {
				code = e.Code
			}

			// Log error ke audit log
			details := map[string]interface{}{
				"status_code": code,
				"error":       err.Error(),
			}
			LogTechnicalError(err, c, details)

			return c.Status(code).JSON(ErrorResponse{
				Error:   "internal_error",
				Message: err.Error(),
			})
		},
	})

	// Middleware global
	app.Use(RecoverMiddleware)                       // Custom recover middleware dengan error logging (harus di awal)
	app.Use(logger.New())                            // Logger middleware
	app.Use(requestid.New())                         // Request ID middleware
	app.Use(SecurityHeadersMiddleware)               // Header keamanan
	app.Use(ErrorHandlerMiddleware)                  // Log error teknis ke audit log
	app.Use(RateLimitMiddleware(generalRateLimiter)) // Rate limiting umum

	// CORS dengan peningkatan keamanan
	app.Use(cors.New(cors.Config{
		AllowOrigins:     "http://localhost:5173,http://localhost:3000",
		AllowMethods:     "GET,POST,PUT,DELETE,OPTIONS,PATCH",
		AllowHeaders:     "Accept,Authorization,Content-Type,X-CSRF-Token,X-Requested-With",
		ExposeHeaders:    "Link,X-Total-Count",
		AllowCredentials: true,
		MaxAge:           300,
	}))

	// Routes
	app.Get("/", indexHandler)
	app.Get("/health", healthHandler)

	// API v1
	api := app.Group("/api/v1")
	api.Get("/", apiInfoHandler)

	// Endpoint token CSRF (public, tidak perlu auth)
	api.Get("/csrf-token", GetCSRFTokenHandler)

	// Route files (public - untuk serve files dari storage, bisa diakses via iframe)
	// Catatan: Route ini public karena iframe tidak bisa mengirim cookie dengan mudah
	// Security: File tetap aman karena path file menggunakan UUID yang tidak bisa ditebak
	api.Get("/files/*", http.ServeFile)

	// Route autentikasi (public) - dengan rate limiting
	// Catatan Teknis:
	// - Endpoint register dihapus karena akan dihandle oleh modul management user
	// - Rate limiting khusus untuk auth endpoints (5 req/min, burst: 5) untuk prevent brute force
	// - Login endpoint mendukung 2FA, jika user memiliki 2FA aktif, akan return requires_2fa: true
	authPublic := api.Group("/auth", AuthRateLimitMiddleware)
	authPublic.Post("/login", Login)

	// Route yang dilindungi (memerlukan JWT)
	// Catatan Teknis:
	// - Semua endpoint di bawah ini memerlukan JWT token valid (dalam cookie atau Authorization header)
	// - CSRF protection diterapkan untuk state-changing methods (POST, PUT, DELETE, PATCH)
	// - GET methods tidak memerlukan CSRF token karena read-only
	protected := api.Group("", JWTAuthMiddleware, CSRFMiddleware)

	// Endpoint profil dan logout user
	protected.Get("/auth/profile", GetProfile)
	protected.Post("/auth/logout", Logout)

	// Route 2FA (Two-Factor Authentication)
	// Catatan: Semua endpoint 2FA memerlukan user sudah login terlebih dahulu
	protected.Post("/auth/2fa/generate", Generate2FASecret) // Generate secret dan QR code
	protected.Post("/auth/2fa/verify", Verify2FA)           // Verifikasi kode dan aktifkan 2FA
	protected.Post("/auth/2fa/disable", Disable2FA)         // Nonaktifkan 2FA (membutuhkan CSRF token)
	protected.Get("/auth/2fa/status", Get2FAStatus)         // Cek status 2FA user

	// Route audit logs (untuk monitoring dan audit trail)
	// Catatan: Admin/superadmin dapat melihat semua logs, user reguler hanya melihat logs sendiri
	protected.Get("/audit-logs", GetAuditLogsHandler)           // List audit logs dengan pagination dan filter
	protected.Get("/audit-logs/stats", GetAuditLogStatsHandler) // Statistik audit logs (total, breakdown, retention policy)

	// Route documents (dilindungi)
	documentHandler := http.NewDocumentHandler(usecase.NewDocumentUseCase())
	protected.Get("/documents/folders", documentHandler.ListFolders)     // List folders
	protected.Post("/documents/folders", documentHandler.CreateFolder)   // Create folder
	protected.Get("/documents", documentHandler.ListDocuments)           // List semua documents
	protected.Get("/documents/summary", documentHandler.DocumentSummary) // Ringkasan dokumen (stats)
	protected.Get("/documents/:id", documentHandler.GetDocument)         // Detail document by ID
	protected.Post("/documents/upload", documentHandler.UploadDocument)  // Upload document baru (memerlukan CSRF)
	protected.Put("/documents/:id", documentHandler.UpdateDocument)      // Update metadata dokumen
	protected.Delete("/documents/:id", documentHandler.DeleteDocument)   // Delete document (memerlukan CSRF)

	// Route financial reports (dilindungi)
	// NOTE: Routes yang lebih spesifik harus diletakkan SEBELUM routes yang lebih umum (dengan :param)
	financialReportHandler := http.NewFinancialReportHandler(usecase.NewFinancialReportUseCase())
	
	// Bulk upload endpoints (harus sebelum /financial-reports/:id karena lebih spesifik)
	protected.Get("/financial-reports/bulk-upload/template", financialReportHandler.GenerateBulkUploadTemplate)    // Download bulk upload template
	protected.Post("/financial-reports/bulk-upload/validate", financialReportHandler.ValidateBulkExcelFile)        // Validate bulk upload Excel file
	protected.Post("/financial-reports/bulk-upload", financialReportHandler.UploadBulkFinancialReports)             // Upload bulk financial reports
	
	// Other specific routes (harus sebelum /financial-reports/:id)
	protected.Get("/financial-reports/company/:company_id", financialReportHandler.GetFinancialReportsByCompanyID) // Get all financial reports for a company
	protected.Get("/financial-reports/compare", financialReportHandler.GetComparison)                            // Get comparison RKAP vs Realisasi YTD
	protected.Get("/financial-reports/rkap-years/:company_id", financialReportHandler.GetRKAPYearsByCompanyID) // Get RKAP years for a company
	
	// General CRUD routes (dengan :id parameter - harus di akhir)
	protected.Post("/financial-reports", financialReportHandler.CreateFinancialReport)                           // Create financial report
	protected.Get("/financial-reports/:id", financialReportHandler.GetFinancialReportByID)                      // Get financial report by ID
	protected.Put("/financial-reports/:id", financialReportHandler.UpdateFinancialReport)                       // Update financial report
	protected.Delete("/financial-reports/:id", financialReportHandler.DeleteFinancialReport)                    // Delete financial report
	
	protected.Get("/companies/:company_id/performance/export/excel", financialReportHandler.ExportPerformanceExcel) // Export performance Excel

	// Swagger
	app.Get("/swagger/*", swagger.HandlerDefault)

	port := os.Getenv("PORT")
	if port == "" {
		port = "8080"
	}

	log.Printf("Server starting on port :%s", port)
	log.Printf("Swagger UI available at http://localhost:%s/swagger/index.html", port)
	log.Printf("API Base URL: http://localhost:%s/api/v1", port)

	if err := app.Listen(":" + port); err != nil {
		log.Fatal(err)
	}
}

// indexHandler handles the root endpoint (untuk Fiber)
// @Summary      Root endpoint
// @Description  Returns API information
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       / [get]
func indexHandler(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"message": "Pedeve App Backend API",
		"version": "1.0.0",
		"swagger": "/swagger/index.html",
	})
}

// healthHandler handles health check (untuk Fiber)
// @Summary      Health check
// @Description  Returns health status
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]string
// @Router       /health [get]
func healthHandler(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"status":  "OK",
		"service": "pedeve-backend",
	})
}

// apiInfoHandler returns API information (untuk Fiber)
// @Summary      API Information
// @Description  Returns API version and endpoints
// @Tags         General
// @Accept       json
// @Produce      json
// @Success      200  {object}  map[string]interface{}
// @Router       /api/v1 [get]
func apiInfoHandler(c *fiber.Ctx) error {
	return c.JSON(fiber.Map{
		"api":     "Pedeve App Backend API",
		"version": "1.0.0",
		"endpoints": fiber.Map{
			"documents": "/api/v1/documents",
			"swagger":   "/swagger/index.html",
		},
	})
}